feat: Include favicon + apple-touch-icon to root
Add the icons to the root directory so non-HTML content
can see it. https://realfavicongenerator.net/favicon_checker
says "There is no Touch icon in the root directory" and
"There is no favicon.ico in the root of your web site"
when checking for favicons.

Based on this answer: https://stackoverflow.com/a/21359390